+++ Apermo/Tests/WordPress/ImplicitPostFunctionUnitTest.php
@@ -23,9 +23,29 @@ class ImplicitPostFunctionUnitTest extends AbstractSniffUnitTest {
 	 */
 	protected function getErrorList( $testFile = '' ) {
 		return [
+			10 => 1,
 			11 => 1,
 			12 => 1,
 			13 => 1,
+			14 => 1,
+			17 => 1,
+			18 => 1,
+			26 => 1,
+			27 => 1,
+			28 => 1,
+			29 => 1,
+			30 => 1,
+			31 => 1,
+			32 => 1,
+			33 => 1,
+			34 => 1,
+			35 => 1,
+			36 => 1,
+			37 => 1,
+			38 => 1,
+			53 => 1,
+			67 => 1,
+			71 => 1,
 		];
 	}
 
@@ -38,15 +58,10 @@ class ImplicitPostFunctionUnitTest extends AbstractSniffUnitTest {
 	 */
 	protected function getWarningList( $testFile = '' ) {
 		return [
-			16 => 1,
-			17 => 1,
-			18 => 1,
 			21 => 1,
 			22 => 1,
-			38 => 1,
-			44 => 1,
-			49 => 1,
-			53 => 1,
+			23 => 1,
+			62 => 1,
 		];
 	}
 }